sign out and int erc721

+++ Web3Unity/Scripts/Prefabs/ERC721/ERC721BalanceOfExample.cs
@@ -12,7 +12,7 @@ public class ERC721BalanceOfExample : MonoBehaviour
         string contract = "0x60f80121c31a0d46b5279700f9df786054aa5ee5";
         string account = "0x6b2be2106a7e883f282e2ea8e203f516ec5b77f7";
 
-        BigInteger balance = await ERC721.BalanceOf(chain, network, contract, account);
+        int balance = await ERC721.BalanceOf(chain, network, contract, account);
         print(balance);
     }
 }
